+++ src/main/kotlin/ee/carlrobert/codegpt/ui/textarea/header/UserInputHeaderPanel.kt
@@ -1,7 +1,6 @@
 package ee.carlrobert.codegpt.ui.textarea.header
 
 import com.intellij.icons.AllIcons
-import com.intellij.openapi.application.runInEdt
 import com.intellij.openapi.components.service
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.editor.EditorKind
@@ -17,6 +16,7 @@ import com.intellij.util.ui.JBUI
 import ee.carlrobert.codegpt.EditorNotifier
 import ee.carlrobert.codegpt.actions.IncludeFilesInContextNotifier
 import ee.carlrobert.codegpt.ui.WrapLayout
+import ee.carlrobert.codegpt.ui.textarea.PromptTextField
 import ee.carlrobert.codegpt.ui.textarea.header.tag.*
 import ee.carlrobert.codegpt.ui.textarea.suggestion.SuggestionsPopupManager
 import ee.carlrobert.codegpt.util.EditorUtil
@@ -32,7 +32,8 @@ import kotlin.math.min
 
 class UserInputHeaderPanel(
     private val project: Project,
-    suggestionsPopupManager: SuggestionsPopupManager
+    suggestionsPopupManager: SuggestionsPopupManager,
+    private val promptTextField: PromptTextField
 ) : JPanel(WrapLayout(FlowLayout.LEFT, 4, 4)), TagManagerListener {
 
     companion object {
@@ -41,7 +42,7 @@ class UserInputHeaderPanel(
     }
 
     private val tagManager = TagManager()
-    private val selectedFileTagPanel = object : SelectedFileTagPanel(project) {
+    private val selectedFileTagPanel = object : SelectedFileTagPanel(project, promptTextField) {
         override fun onClose() {
             this.isVisible = false
             if (tagManager.getTags().isEmpty()) {
@@ -55,7 +56,7 @@ class UserInputHeaderPanel(
         border = JBUI.Borders.empty(3, 4)
         isVisible = getSelectedEditor(project) == null
     }
-    private val selectionTagPanel = SelectionTagPanel(project)
+    private val selectionTagPanel = SelectionTagPanel(project, promptTextField)
     private val defaultHeaderTagsPanel = CustomFlowPanel().apply {
         add(AddButton {
             if (suggestionsPopupManager.isPopupVisible()) {
@@ -76,7 +77,7 @@ class UserInputHeaderPanel(
     }
 
     fun getSelectedTags(): List<TagDetails> {
-        val selectedTags = tagManager.getSelectedTags().toMutableList()
+        val selectedTags = tagManager.getTags().filter { it.selected }.toMutableList()
 
         val selectedFile = getSelectedFile()
         if (selectedFileTagPanel.isVisible && selectedFileTagPanel.tagDetails.selected && selectedFile != null) {
@@ -96,18 +97,18 @@ class UserInputHeaderPanel(
 
     override fun onTagAdded(tag: TagDetails) {
         emptyText.isVisible = false
-        add(createTag(tag), getNextInsertionIndex())
+        add(createTag(tag), getInsertionIndex())
+
+        if (tagManager.getTags().filter { !it.selected }.size > 2) {
+            components
+                .lastOrNull { it is TagPanel && it.tagDetails is FileTagDetails && !it.isSelected }
+                ?.let { tagManager.removeTag((it as TagPanel).tagDetails.id) }
+        }
+
+        promptTextField.requestFocus()
+
         revalidate()
         repaint()
-    }
-
-    private fun getNextInsertionIndex(): Int {
-        val lastSelectionTagIndex = getLastSelectedTagIndex()
-        return if (lastSelectionTagIndex != -1) {
-            min(lastSelectionTagIndex + TAG_INSERTION_OFFSET + 1, components.size)
-        } else {
-            TAG_INSERTION_OFFSET
-        }
     }
 
     override fun onTagRemoved(tag: TagDetails) {
@@ -122,7 +123,7 @@ class UserInputHeaderPanel(
         if (tagManager.getTags().isEmpty() && !selectedFileTagPanel.isVisible) {
             emptyText.isVisible = true
         }
-
+        promptTextField.requestFocus()
         revalidate()
         repaint()
     }
@@ -136,7 +137,13 @@ class UserInputHeaderPanel(
     }
 
     private fun createTag(tagDetails: TagDetails) =
-        object : TagPanel(tagDetails, tagDetails is FileTagDetails) {
+        object : TagPanel(tagDetails, true) {
+
+            init {
+                cursor =
+                    if (tagDetails is FileTagDetails) Cursor(Cursor.HAND_CURSOR) else Cursor(Cursor.DEFAULT_CURSOR)
+            }
+
             override fun onSelect(tagDetails: TagDetails) {
                 if (tagDetails is FileTagDetails) {
                     if (tagDetails.selected) {
@@ -148,7 +155,7 @@ class UserInputHeaderPanel(
 
                     val canAddNewTag = tagManager.getTags()
                         .filterIsInstance<FileTagDetails>()
-                        .count { !it.selected } <= 2
+                        .count { !it.selected } < 2
                     if (canAddNewTag) {
                         addNextOpenFile()
                     }
@@ -163,9 +170,18 @@ class UserInputHeaderPanel(
 
     private fun updateTagPosition(tag: TagPanel) {
         remove(tag)
-        add(tag, getNextInsertionIndex())
+        add(tag, getInsertionIndex())
         revalidate()
         repaint()
+    }
+
+    private fun getInsertionIndex(): Int {
+        val lastSelectionTagIndex = getLastSelectedTagIndex()
+        return if (lastSelectionTagIndex != -1) {
+            min(lastSelectionTagIndex + TAG_INSERTION_OFFSET + 1, components.size)
+        } else {
+            TAG_INSERTION_OFFSET
+        }
     }
 
     private fun getLastSelectedTagIndex(): Int =
@@ -209,7 +225,6 @@ class UserInputHeaderPanel(
     private fun initializeEventListeners() {
         project.messageBus.connect().apply {
             subscribe(EditorNotifier.SelectionChange.TOPIC, EditorSelectionChangeListener())
-            subscribe(EditorNotifier.Created.TOPIC, EditorCreatedListener())
             subscribe(EditorNotifier.Released.TOPIC, EditorReleasedListener())
             subscribe(FileEditorManagerListener.FILE_EDITOR_MANAGER, FileSelectionListener())
             subscribe(
@@ -255,18 +270,6 @@ class UserInputHeaderPanel(
         }
     }
 
-    private inner class EditorCreatedListener : EditorNotifier.Created {
-        override fun editorCreated(editor: Editor) {
-            editor.virtualFile?.let { editorFile ->
-                if (selectedFileTagPanel.isVisible && getSelectedFile() == editorFile) {
-                    return
-                }
-
-                tagManager.addTag(FileTagDetails(editorFile, false))
-            }
-        }
-    }
-
     private inner class EditorReleasedListener : EditorNotifier.Released {
         override fun editorReleased(editor: Editor) {
             if (editor.editorKind == EditorKind.MAIN_EDITOR && !editor.isDisposed && editor.virtualFile != null) {
@@ -281,22 +284,20 @@ class UserInputHeaderPanel(
 
     private inner class FileSelectionListener : FileEditorManagerListener {
         override fun selectionChanged(event: FileEditorManagerEvent) {
-            val currentTagDetails = (selectedFileTagPanel.tagDetails as? FileTagDetails) ?: return
             event.newFile?.let { newFile ->
-                if (event.oldFile == currentTagDetails.virtualFile) {
-                    tagManager.removeFileTag(newFile)
-                    tagManager.addTag(currentTagDetails)
+                var existingFileTag = tagManager.getFileTag(newFile)
+                if (existingFileTag != null) {
+                    tagManager.removeTag(existingFileTag.id)
+                } else {
+                    existingFileTag = FileTagDetails(newFile).apply { selected = false }
                 }
 
-                val existingTag = tagManager.getFileTag(newFile)
-                runInEdt {
-                    if (existingTag == null) {
-                        selectedFileTagPanel.update(FileTagDetails(newFile))
-                    } else {
-                        selectedFileTagPanel.update(existingTag)
-                    }
-                    emptyText.isVisible = false
+                if (selectedFileTagPanel.tagDetails !is EmptyTagDetails) {
+                    tagManager.addTag(selectedFileTagPanel.tagDetails)
                 }
+
+                selectedFileTagPanel.update(existingFileTag)
+                emptyText.isVisible = false
             }
         }
     }